requirements.txt
alembic==1.11.1
apispec==6.3.0
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Migrate==4.0.4
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
webargs==8.3.0
Werkzeug==2.3.6
